refactor(pci-instances): restructure types folder
ref:1864
