--- packages/manager/apps/pci-instance/.eslintrc.cjs
@@ -1,24 +1,0 @@
-module.exports = {
-  extends: [
-    '../../../../.eslintrc.js',
-    'plugin:react/recommended',
-    'plugin:react/jsx-runtime',
-    'plugin:react-hooks/recommended',
-    'plugin:prettier/recommended',
-  ],
-  settings: {
-    react: {
-      version: 'detect',
-    },
-  },
-  globals: {
-    __VERSION__: true,
-  },
-
-  rules: {
-    '@typescript-eslint/no-unused-vars': [
-      'error',
-      { ignoreRestSiblings: true },
-    ],
-  },
-};
