--- packages/manager/apps/pci-instance/README.md
@@ -1,1 +1,0 @@
-# PCI Instance
