--- packages/manager/apps/pci-instance/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>OVHcloud</title>
-    </head>
-    <body>
-        <div id="root"></div>
-        <script type="module" src="./src/main.tsx"></script>
-    </body>
-</html>
