--- packages/manager/apps/pci-instance/package.json
@@ -1,70 +1,0 @@
-{
-  "name": "@ovh-ux/manager-pci-instance-app",
-  "version": "0.0.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "build": "tsc --project tsconfig.build.json && vite build",
-    "coverage": "vitest run --coverage",
-    "dev": "vite",
-    "lint": "eslint ./src",
-    "start": "lerna exec --stream --scope='@ovh-ux/manager-pci-instance-app' --include-dependencies -- npm run build --if-present",
-    "start:dev": "lerna exec --stream --scope='@ovh-ux/manager-pci-instance-app' --include-dependencies -- npm run dev --if-present",
-    "start:watch": "lerna exec --stream --parallel --scope='@ovh-ux/manager-pci-instance-app' --include-dependencies -- npm run dev:watch --if-present",
-    "test": "vitest run",
-    "test:watch": "vitest --watch",
-    "type:check": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@ovh-ux/manager-config": "^7.3.2",
-    "@ovh-ux/manager-core-api": "^0.8.0",
-    "@ovh-ux/manager-react-core-application": "^0.10.0",
-    "@ovh-ux/manager-react-shell-client": "^0.7.0",
-    "@ovh-ux/manager-tailwind-config": "^0.2.0",
-    "@ovhcloud/manager-components": "^1.26.0",
-    "@ovhcloud/ods-common-core": "17.2.2",
-    "@ovhcloud/ods-common-stencil": "17.2.2",
-    "@ovhcloud/ods-common-theming": "17.2.2",
-    "@ovhcloud/ods-components": "17.2.2",
-    "@ovhcloud/ods-theme-blue-jeans": "17.2.2",
-    "@tanstack/react-query": "^5.51.21",
-    "axios": "^1.6.8",
-    "element-internals-polyfill": "^1.3.10",
-    "i18next": "^23.8.2",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-i18next": "^14.0.5",
-    "react-router-dom": "^6.3.0",
-    "zustand": "^4.5.0"
-  },
-  "devDependencies": {
-    "@jest/globals": "^29.7.0",
-    "@ovh-ux/manager-vite-config": "^0.8.0",
-    "@tanstack/react-query-devtools": "^5.51.21",
-    "@testing-library/dom": "^9.3.4",
-    "@testing-library/jest-dom": "^6.4.2",
-    "@testing-library/react": "^14.2.1",
-    "@testing-library/user-event": "^14.5.2",
-    "@types/jest": "^29.5.12",
-    "@types/react": "^18.2.55",
-    "@types/react-dom": "^18.2.19",
-    "@types/react-query": "^1.2.9",
-    "@vitejs/plugin-react": "^4.2.1",
-    "@vitest/coverage-v8": "^1.2.2",
-    "autoprefixer": "^10.4.17",
-    "eslint": "^8.57.0",
-    "eslint-plugin-react": "^7.35.0",
-    "postcss": "^8.4.35",
-    "tailwindcss": "^3.4.4",
-    "vite": "^5.2.13",
-    "vitest": "^1.2.0"
-  },
-  "regions": [
-    "CA",
-    "EU",
-    "US"
-  ],
-  "universes": [
-    "@ovh-ux/manager-public-cloud"
-  ]
-}
