+++ packages/manager/apps/pci-instances/package.json
@@ -18,7 +18,7 @@
   "dependencies": {
     "@ovh-ux/manager-config": "^7.3.2",
     "@ovh-ux/manager-core-api": "^0.8.0",
-    "@ovh-ux/manager-pci-common": "^0.4.1",
+    "@ovh-ux/manager-pci-common": "^0.6.1",
     "@ovh-ux/manager-react-components": "^1.26.0",
     "@ovh-ux/manager-react-core-application": "^0.10.0",
     "@ovh-ux/manager-react-shell-client": "^0.7.0",
